requirements.txt
google-api-python-client>=2.127.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
transformers>=4.41.1
torch>=2.2.2
beautifulsoup4>=4.12.3
